requirements.txt
flask==3.0.3
flask_cors==5.0.0
flask_restx==1.3.0
flask_sqlalchemy==3.1.1
python_dotenv==1.0.1
SQLAlchemy==2.0.35
werkzeug==3.0.4